Initial upload of "subtle_wip" colorscheme.
I have been using this colorscheme for ~6 years now, with little change. I use
it mainly on a tmux with proper 256 color support. I use it with dark background.

I recently made some fixes, to have it work on NeoVim, also adjust the color of
the pop-up menu, which didn't have enough contrast on dark background, made
hlsearch and matchit a lot more subtle, and also some other minor fixes. (Most
of the fixes went into dark background 256 color which, again, is the
combination I use myself.)

While at it, I decided to turn it into a separate plug-in that I can manage
with vim-plug, so that I can more easily have it available on all my hosts.
This commit makes it into a plug-in. :-)